split up massive spec into still pretty big files
